[DDW-321] Separates Wallets requests into their own sub-directory
