+++ source/renderer/app/api/wallets/requests/changeSpendingPassword.js
@@ -1,12 +1,13 @@
 // @flow
-import type { AdaWallet, RequestConfig } from './types';
+import type { RequestConfig } from '../../common/types';
+import type { AdaWallet } from '../types';
 import { encryptPassphrase } from '../../utils';
 import { request } from '../../utils/request';
 
 export type ChangeSpendingPasswordParams = {
   walletId: string,
-  oldPassword?: string,
-  newPassword: string,
+  oldPassword: ?string,
+  newPassword: ?string,
 };
 
 export const changeSpendingPassword = (
