+++ source/renderer/app/api/wallets/requests/createWallet.js
@@ -1,14 +1,19 @@
 // @flow
-import type { AdaWallet, AdaWalletInitData, RequestConfig } from './types';
+import type { RequestConfig } from '../../common/types';
+import type { AdaWallet, WalletAssuranceLevel } from '../types';
 import { request } from '../../utils/request';
 
-export type CreateWalletParams = {
-  walletInitData: AdaWalletInitData
+export type WalletInitData = {
+  operation: 'create' | 'restore',
+  backupPhrase: [string],
+  assuranceLevel: WalletAssuranceLevel,
+  name: string,
+  spendingPassword: ?string,
 };
 
 export const createWallet = (
   config: RequestConfig,
-  { walletInitData }: CreateWalletParams
+  { walletInitData }: { walletInitData: WalletInitData }
 ): Promise<AdaWallet> => (
   request({
     hostname: 'localhost',
