+++ source/renderer/app/api/wallets/requests/resetWalletState.js
@@ -1,6 +1,6 @@
 // @flow
+import type { RequestConfig } from '../../common/types';
 import { request } from '../../utils/request';
-import type { RequestConfig } from './types';
 
 export const resetWalletState = (
   config: RequestConfig
